make follow && install vuetify && edit upload component

+++ database/migrations/2020_05_28_084537_create_friends.php
@@ -15,8 +15,8 @@ class CreateFriends extends Migration
     {
         Schema::create('friends', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('friend_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('friend_id');
             $table->timestamps();
         });
     }
